common: Create a common package to share functions
There is some overlap between cloud and weldr's common functions, this
moves them into an internal/common package so they can be shared.

This commit creates the shared package and refactors weldr to use them.

Related: RHEL-60148
